% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uibe.r
\docType{data}
\name{apis}
\alias{apis}
\title{Qutke Data API}
\format{\preformatted{List of 7
 $ tradingDay    : chr "http://test-nquote.qutke.com/tradingDay"
 $ keyMap        : chr "http://test-nquote.qutke.com/keyMap"
 $ mktDaily      : chr "http://test-nquote.qutke.com/mktDaily"
 $ mktFwdDaily   : chr "http://test-nquote.qutke.com/mktFwdDaily"
 $ mktDataIndex  : chr "http://test-nquote.qutke.com/mktDataIndex"
 $ industryType  : chr "http://test-nquote.qutke.com/industryType"
 $ financialIndex: chr "http://test-nquote.qutke.com/financialIndex"
}}
\usage{
apis
}
\description{
Qutke Data API
}
\keyword{datasets}

