% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uibe.r
\name{getData}
\alias{getData}
\title{Get data from HTTP}
\usage{
getData(name, args = NULL, key = NULL)
}
\arguments{
\item{name}{character}

\item{args}{list}

\item{key}{character}
}
\value{
data.frame
}
\description{
Get data from HTTP
}
\examples{
\dontrun{
TradingDay<-getData('TradingDay',key='xxxx')

KeyMap1<-getData('KeyMap',args=list(qtid='000001.SZ'),key='xxxx')

stock1<-getData('stock',args=list(qtid='000001.SZ'),key='xxxx')
stock2<-getData('stock',args=list(date='2015-09-29'),key='xxxx')
stock3<-getData('stock',args=list(qtid='000001.SZ',date='2015-09-29'),key='xxxx')

stockF1<-getData('stockFwd',args=list(qtid='000001.SZ'),key='xxxx')
stockF2<-getData('stockFwd',args=list(date='2015-09-29'),key='xxxx')
stockF3<-getData('stockFwd',args=list(qtid='000001.SZ',date='2015-09-29'),key='xxxx')

index1<-getData('index',args=list(qtid='000001.SH'),key='xxxx')
index2<-getData('index',args=list(date='2015-09-29'),key='xxxx')
index3<-getData('index',args=list(qtid='000001.SH',date='2015-09-29'),key='xxxx')

industry0<-getData('industry',key='xxxx')
industry1<-getData('industry',args=list(qtid='000001.SZ'),key='xxxx')
industry2<-getData('industry',args=list(qtid='000001.SZ',sw1='sw1'),key='xxxx')
industry3<-getData('industry',args=list(qtid='000001.SZ',sw1='sw1',sw2='sw2'),key='xxxx')
industry4<-getData('industry',args=list(qtid='000001.SZ',sw1='sw1',sw2='sw2',sw3='sw3'),key='xxxx')

fund1<-getData('fund',args=list(qtid='000001.SZ'),key='xxxx')
fund2<-getData('fund',args=list(date='2015-09-29'),key='xxxx')
fund3<-getData('fund',args=list(qtid='000001.SZ',date='2015-09-29'),key='xxxx')
}
}
\author{
Dan Zhang
}

