% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uibe.r
\docType{data}
\name{host}
\alias{host}
\title{Qutke host server}
\format{\preformatted{List of 2
 $ host1: chr "http://datas.qutke.com"
 $ host2: chr "http://uibe.qutke.com/api/qutke_inner_api_data"
}}
\usage{
host
}
\description{
Qutke host server
}
\keyword{datasets}

