% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uibe.r
\docType{data}
\name{apis}
\alias{apis}
\title{Qutke Data API}
\format{\preformatted{List of 7
 $ TradingDay: chr "http://test-nquote.qutke.com/tradingDay"
 $ KeyMap    : chr "http://test-nquote.qutke.com/keymap"
 $ stock     : chr "http://test-nquote.qutke.com/mktData"
 $ stockFwd  : chr "http://test-nquote.qutke.com/mktFwdDaily"
 $ index     : chr "http://test-nquote.qutke.com/mktDataIndex"
 $ industry  : chr "http://test-nquote.qutke.com/industry"
 $ fund      : chr "http://test-nquote.qutke.com/fund"
}}
\usage{
apis
}
\description{
Qutke Data API
}
\keyword{datasets}

