% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uibe.r
\docType{data}
\name{apis}
\alias{apis}
\title{Qutke Data API}
\format{\preformatted{List of 8
 $ tradingDay      : chr "http://localhost:8090/api/qutkedata/tradingDay"
 $ keyMap          : chr "http://localhost:8090/api/qutkedata/keyMap"
 $ mktDaily        : chr "http://localhost:8090/api/qutkedata/mktDaily"
 $ mktFwdDaily     : chr "http://localhost:8090/api/qutkedata/mktFwdDaily"
 $ mktDataIndex    : chr "http://localhost:8090/api/qutkedata/mktDataIndex"
 $ industryType    : chr "http://localhost:8090/api/qutkedata/industryType"
 $ financialIndex  : chr "http://localhost:8090/api/qutkedata/financialIndex"
 $ securitiesMargin: chr "http://localhost:8090/api/qutkedata/securitiesMargin"
}}
\usage{
apis
}
\description{
Qutke Data API
}
\keyword{datasets}

