% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uibe.r
\docType{data}
\name{apis}
\alias{apis}
\title{Qutke Data API}
\format{\preformatted{List of 8
 $ tradingDay      : chr "http://datas.qutke.com/tradingDay"
 $ keyMap          : chr "http://datas.qutke.com/keyMap"
 $ mktDaily        : chr "http://datas.qutke.com/mktDaily"
 $ mktFwdDaily     : chr "http://datas.qutke.com/mktFwdDaily"
 $ mktDataIndex    : chr "http://datas.qutke.com/mktDataIndex"
 $ industryType    : chr "http://datas.qutke.com/industryType"
 $ financialIndex  : chr "http://datas.qutke.com/financialIndex"
 $ securitiesMargin: chr "http://datas.qutke.com/securitiesMargin"
}}
\usage{
apis
}
\description{
Qutke Data API
}
\keyword{datasets}

