% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uibe.r
\docType{data}
\name{host}
\alias{host}
\title{Qutke host server}
\format{\preformatted{List of 2
 $ host1: chr "http://localhost:8090/api/qutkedata"
 $ host2: chr "http://localhost:8090/api/adduserdata"
}}
\usage{
host
}
\description{
Qutke host server
}
\keyword{datasets}

